pyproject.toml
[project]
name = "xcoder"
version = "0.1.0"
description = ""
authors = [{ name = "Danila", email = "crowo4ka@gmail.com" }]
requires-python = ">=3.11,<4"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "sc-compression==0.6.6",
    "colorama==0.4.6",
    "zstandard>=0.23.0,<0.24",
    "pillow~=11.2.1",
    "loguru==0.7.3",
]

[project.optional-dependencies]
lzham = ["pylzham>=0.1.3,<0.2"]

[dependency-groups]
dev = [
    "pyright>=1.1.376,<2",
    "black>=24.8.0,<25",
    "pre-commit>=3.8.0,<4",
    "ruff>=0.11.2,<0.12",
]

[project.scripts]
xcoder = "xcoder.__main__:main"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[tool.uv]

[tool.uv.build-backend]
module-name = "xcoder"

[tool.isort]
profile = "black"
force_sort_within_sections = true  # Don't group `from` imports separately
order_by_type = true  # Order by CONSTANT, CamelCase, snake_case

